🚚 Move and rename methods component
