+++ apps/cookbook/src/app/showcase/dropdown-showcase/dropdown-showcase.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { ShowcaseEvent } from '~/app/shared/api-description/api-description-events/api-description-events.component';
-import { ShowcaseMethod } from '~/app/shared/showcase-methods/showcase-methods.component';
+import { ShowcaseMethod } from '~/app/shared/api-description/api-description-methods/api-description-methods.component';
 import { ShowcaseProperty } from '~/app/shared/showcase-properties/showcase-properties.component';
 
 @Component({
