+++ apps/cookbook/src/app/showcase/showcase.module.ts
@@ -7,9 +7,9 @@ import { KirbyModule } from '@kirbydesign/designsystem';
 
 import { IphoneModule } from '../iphone/iphone.module';
 import { ShowcaseEventsComponent } from '../shared/api-description/api-description-events/api-description-events.component';
+import { ShowcaseMethodsComponent } from '../shared/api-description/api-description-methods/api-description-methods.component';
 import { CodeViewerComponent } from '../shared/code-viewer/code-viewer.component';
 import { ExampleViewerComponent } from '../shared/example-viewer/example-viewer.component';
-import { ShowcaseMethodsComponent } from '../shared/showcase-methods/showcase-methods.component';
 import { ShowcasePropertiesComponent } from '../shared/showcase-properties/showcase-properties.component';
 
 import { DividerShowcaseComponent } from './divider-showcase/divider-showcase.component';
